разработка ограничения пользователя

+++ src/eos-user-params/rights-delo/rights-delo-absolute-rights/absolute-rights-classif/absolute-rights-classif-node.ts
@@ -124,7 +124,9 @@ export class RightClassifNode {
         this.isShell = true;
         this._component.addInstance(this._config, this)
             .then(data => {
-                this._item.label = 'Пользователи (доступ ограничен)';
+                if (this._config.rootLabel === 'Центральная картотека') {
+                    this._item.label = 'Пользователи (доступ ограничен)';
+                }
                 const newList: NodeDocsTree[] = [];
                 if (data) {
                     data.forEach(entity => {
@@ -201,22 +203,62 @@ export class RightClassifNode {
     }
     DeleteInstance() {
         if (this.curentSelectedNode) {
-            this.listContent = this.listContent.filter(node => node !== this.curentSelectedNode);
-            this._parentNode.pushChange({
-                method: 'DELETE',
-                due: this.curentSelectedNode.DUE,
-                funcNum: this.key,
-                data: this.curentSelectedNode.data['userTech']
-            });
-            if (this.listContent.length === 0) {
-                this._item.label =  'Пользователи';
+            if (this.curentSelectedNode.children.length !== 0 && this.curentSelectedNode.DUE !== '0.' && this._config.rootLabel === 'Центральная картотека') {
+                this._component.strNewCards = [];
+                this._component.strNewCards = [{value: 'Исключить из перечня подчиненные картотеки:'}];
+                this.curentSelectedNode.children.forEach((card) => {
+                    this._component.strNewCards.push({value: String(card.label), due: card.DUE});
+                });
+                this._component._userParmSrv.confirmCallCard(this._component.newCards).then((answer) => {
+                    if (answer) {
+                        const childDue = this.curentSelectedNode.children.map(item => {
+                            this._parentNode.pushChange({
+                                method: 'DELETE',
+                                due: item.DUE,
+                                funcNum: 1,
+                                data: item.data['userTech']
+                            });
+                            return item.DUE;
+                        });
+                        this.listContent = this.listContent.filter(node => node !== this.curentSelectedNode && childDue.indexOf(node.DUE) === -1);
+                        this._listUserTech = this._listUserTech.filter(node => childDue.indexOf(node['DUE']) === -1);
+                        this._component.userTechList = this._component.userTechList.filter(node => childDue.indexOf(node['DUE']) === -1);
+                    }
+                    this.listContent = this.listContent.filter(node => node !== this.curentSelectedNode);
+                    this._parentNode.pushChange({
+                        method: 'DELETE',
+                        due: this.curentSelectedNode.DUE,
+                        funcNum: this.key,
+                        data: this.curentSelectedNode.data['userTech']
+                    });
+                    if (this.listContent.length === 0) {
+                        this._item.label =  'Пользователи';
+                    }
+                    const index = this._listUserTech.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
+                    this._listUserTech.splice(index, 1);
+                    const index2 = this._component.userTechList.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
+                    this._component.userTechList.splice(index2, 1);
+                    this.curentSelectedNode = null;
+                    this._component.Changed.emit();
+                });
+            } else {
+                this.listContent = this.listContent.filter(node => node !== this.curentSelectedNode);
+                this._parentNode.pushChange({
+                    method: 'DELETE',
+                    due: this.curentSelectedNode.DUE,
+                    funcNum: this.key,
+                    data: this.curentSelectedNode.data['userTech']
+                });
+                if (this.listContent.length === 0) {
+                    this._item.label =  'Пользователи';
+                }
+                const index = this._listUserTech.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
+                this._listUserTech.splice(index, 1);
+                const index2 = this._component.userTechList.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
+                this._component.userTechList.splice(index2, 1);
+                this.curentSelectedNode = null;
+                this._component.Changed.emit();
             }
-            const index = this._listUserTech.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
-            this._listUserTech.splice(index, 1);
-            const index2 = this._component.userTechList.findIndex(node => this.curentSelectedNode.DUE === node['DUE']);
-            this._component.userTechList.splice(index2, 1);
-            this.curentSelectedNode = null;
-            this._component.Changed.emit();
         }
     }
     select(node: NodeDocsTree) {
